requirements.txt
mysql-connector-python==9.3.0
pymongo==4.12.1
redis==5.2.1